requirements.txt
asgiref==3.8.1
Django==5.1.7
django-crispy-forms==2.3
django-filter==25.1
djangorestframework==3.15.2
isort==6.0.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
